feat: rename in cebabCase
