+++ src/app/App.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { HashRouter, Navigate, Route, Routes } from 'react-router-dom';
-import { Content } from 'src/app/logic/content/Content';
-import { Layout } from 'src/app/logic/layout/Layout';
-import { MainLayout } from 'src/app/logic/main/MainLayout';
-import { BOOK_PAGE_URL,BookPage } from 'src/app/logic/pages/book/BookPage';
-import { Contract, CONTRACT_PAGE_URL } from 'src/app/logic/pages/contract/Contract';
-import { Terms, TERMS_PAGE_URL } from 'src/app/logic/pages/terms/Terms';
-import { BookProvider } from 'src/app/logic/Store';
+import { Content } from 'src/app/logic/content/content';
+import { Layout } from 'src/app/logic/layout/layout';
+import { MainLayout } from 'src/app/logic/main/main-layout';
+import { BOOK_PAGE_URL,BookPage } from 'src/app/logic/pages/book/book-page';
+import { Contract, CONTRACT_PAGE_URL } from 'src/app/logic/pages/contract/contract';
+import { Terms, TERMS_PAGE_URL } from 'src/app/logic/pages/terms/terms';
+import { BookProvider } from 'src/app/logic/store';
 
 /**
  * The main component in app
