+++ src/app/logic/pages/book/book-page.tsx
@@ -2,13 +2,13 @@ import React from 'react';
 import { NavLink, useParams } from 'react-router-dom';
 import clsx from 'clsx';
 import BooksDTO from 'src/app/Books.json';
-import { Button } from 'src/app/components/button/Button';
-import { Header } from 'src/app/components/header/Header';
-import { MenuListInterface } from 'src/app/components/menu/Menu';
-import { Rating } from 'src/app/components/rating/Rating';
-import { Review } from 'src/app/components/review/Review';
-import { BooksInterface } from 'src/app/logic/content/Content';
-import { Footer } from 'src/app/logic/footer/Footer';
+import { Button } from 'src/app/components/button/button';
+import { Header } from 'src/app/components/header/header';
+import { MenuListInterface } from 'src/app/components/menu/menu';
+import { Rating } from 'src/app/components/rating/rating';
+import { Review } from 'src/app/components/review/review';
+import { BooksInterface } from 'src/app/logic/content/content';
+import { Footer } from 'src/app/logic/footer/footer';
 import MenuListDTO from 'src/app/MenuList.json';
 import bookUrl from 'src/resources/book.jpg';
 import bookNotUrl from 'src/resources/book-not.svg';
