+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {createRoot} from 'react-dom/client';
-import {App} from 'src/app/App';
+import {App} from 'src/app/app';
 
 import 'src/index.scss';
